Moved game world palette updating to SceneManager.
